added a bunch of hunav sim nodes and their respective arena node

+++ arena_hunav_sim_bridge/hunav_sim_node_wrapper/stop_and_wait_timer_action.py
@@ -17,7 +17,7 @@ class StopAndWaitTimerAction(BaseNode):
                 name="agent_id",
                 type="int",
             ),
-            InputPort(name="wait_duration", type="float"),
+            InputPort(name="wait_duration", type="double"),
         ]
 
         actions = [Action(ID="StopAndWaitTimerAction", input_ports=input_ports)]
